+++ index.html
@@ -10,6 +10,8 @@
   <script src="https://cdn.jsdelivr.net/npm/bootstrap@5.0.0-beta2/dist/js/bootstrap.bundle.min.js" integrity="sha384-b5kHyXgcpbZJO/tY9Ul7kGkf1S0CWuKcCD38l8YkeH8z8QjE0GmW1gYU5S9FOnJ0" crossorigin="anonymous"></script>
   <link rel="stylesheet" href="css/style.css">
 
+  <script type="module" src="js/dom.js" defer> </script>
+
   <script type="text/javascript" src="js/index.js" defer> </script>
 
 </head>
@@ -26,7 +28,7 @@
         <input class="form-control" name="author" id="author" type="text" placeholder="Author" required>
       </div>
       <div class="form-group">
-        <input class="form-control" type="number" name="pages" id="pages" placeholder="Pages" required>
+        <input class="form-control" min = "0" type="number" name="pages" id="pages" placeholder="Pages" required>
       </div>
       <div class="form-group">
         <p>Have you read this?</p>
@@ -40,7 +42,7 @@
 
       <button id="form-button" type="submit" class="btn btn-primary mb-3">Add Book</button> <br />
     </form>
-    <div id="book-group" class="card-group p-2">
+    <div id="book-group" class="d-flex flex-wrap p-1">
     </div>
   </div>
