+++ package.json
@@ -4,5 +4,22 @@
     "eslint": "6.8.x",
     "eslint-config-airbnb-base": "14.1.x",
     "eslint-plugin-import": "2.20.x"
-  }
+  },
+  "name": "webpack",
+  "description": "This is a mini-project that allows you to",
+  "version": "1.0.0",
+  "main": "index.js",
+  "scripts": {
+    "test": "echo \"Error: no test specified\" && exit 1"
+  },
+  "repository": {
+    "type": "git",
+    "url": "git+https://github.com/newhorizon-tech/library.git"
+  },
+  "author": "",
+  "license": "ISC",
+  "bugs": {
+    "url": "https://github.com/newhorizon-tech/library/issues"
+  },
+  "homepage": "https://github.com/newhorizon-tech/library#readme"
 }
